+++ examples/fdp/seirs_sim.jl
@@ -8,7 +8,7 @@ using DataFrames
 using Plots
 
 # Initialise code run
-handle = initialise()
+handle = DataPipeline.initialise()
 
 # Read model parameters
 path = link_read!(handle, "SEIRS_model/parameters")
@@ -41,4 +41,4 @@ path = link_write!(handle, "figure")
 savefig(g, path)
 
 # Register code run in local registry
-finalise(handle)
+DataPipeline.finalise(handle)